software/bms_gui_bq76925_atmega328p_6s2_20a/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.2
pyserial==3.5
